# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Microfluidic API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/v1/import/dxf": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Import Dxf",
        "operationId": "import_dxf_api_v1_import_dxf_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_dxf_api_v1_import_dxf_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DXFImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_import_dxf_api_v1_import_dxf_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "default_width": {
            "type": "number",
            "title": "Default Width",
            "default": 10.0
          },
          "scale": {
            "type": "number",
            "title": "Scale",
            "default": 1.0
          },
          "layer_widths": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layer Widths"
          },
          "layer_types": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layer Types"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_dxf_api_v1_import_dxf_post"
      },
...
      "DXFImportResponse": {
        "properties": {
          "shapes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Shapes"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "default": []
          },
          "layers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Layers",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "shapes"
        ],
        "title": "DXFImportResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}